WasteAnnotator/pyproject.toml
[tool.poetry]
name = "WasteAnnotator"
version = "0.4.0"
description = ""
authors = ["Cezar Sas", "Matthijs", "Andro"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.104.1"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0"
pandas = "^2.1.3"
setuptools = "^69.0.0"
pydantic = "^2.5.2"
gitpython = "^3.1.40"
loguru = "^0.7.1"
tqdm = "^4.66.1"
more-itertools = "^10.1.0"
sqlalchemy = "^2.0.21"
psycopg = {extras = ["binary"], version = "^3.1.14"}
streamlit = "^1.29.0"
streamlit-extras = "^0.4.7"
scipy = "^1.11.4"
cdlib = "^0.4.0"
wurlitzer = "^3.0.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.16.0"
hydra-core = "^1.3.2"
bayanpy = "^0.7.7"
graph-tools = "^1.14"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = "^0.24.0"
mkdocs-gen-files = "^0.5.0"
mkdocstrings-python = "^1.7.5"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
mkdocs-material = "^9.4.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
